+++ src/System.Windows.Forms.Primitives/src/Interop/WinFormsComWrappers.cs
@@ -16,7 +16,7 @@ internal partial class Interop
     ///
     /// Supports IStream COM interface.
     /// </summary>
-    internal unsafe class WinFormsComWrappers : ComWrappers
+    internal unsafe partial class WinFormsComWrappers : ComWrappers
     {
         private const int S_OK = (int)Interop.HRESULT.S_OK;
         private static readonly ComInterfaceEntry* s_wrapperEntry = InitializeComInterfaceEntry();
@@ -29,9 +29,9 @@ internal partial class Interop
 
         private static ComInterfaceEntry* InitializeComInterfaceEntry()
         {
-            GetIUnknownImpl(out IntPtr fpQueryInteface, out IntPtr fpAddRef, out IntPtr fpRelease);
+            GetIUnknownImpl(out IntPtr fpQueryInterface, out IntPtr fpAddRef, out IntPtr fpRelease);
 
-            IntPtr iStreamVtbl = IStreamVtbl.Create(fpQueryInteface, fpAddRef, fpRelease);
+            IntPtr iStreamVtbl = IStreamVtbl.Create(fpQueryInterface, fpAddRef, fpRelease);
 
             ComInterfaceEntry* wrapperEntry = (ComInterfaceEntry*)RuntimeHelpers.AllocateTypeAssociatedMemory(typeof(WinFormsComWrappers), sizeof(ComInterfaceEntry));
             wrapperEntry->IID = IID_IStream;
@@ -66,347 +66,6 @@ internal partial class Interop
         protected override void ReleaseObjects(IEnumerable objects)
         {
             throw new NotImplementedException();
-        }
-
-        internal static class IStreamVtbl
-        {
-            public static IntPtr Create(IntPtr fpQueryInteface, IntPtr fpAddRef, IntPtr fpRelease)
-            {
-                IntPtr* vtblRaw = (IntPtr*)RuntimeHelpers.AllocateTypeAssociatedMemory(typeof(IStreamVtbl), IntPtr.Size * 14);
-                vtblRaw[0] = fpQueryInteface;
-                vtblRaw[1] = fpAddRef;
-                vtblRaw[2] = fpRelease;
-                vtblRaw[3] = (IntPtr)(delegate* unmanaged<IntPtr, byte*, uint, uint*, int>)&Read;
-                vtblRaw[4] = (IntPtr)(delegate* unmanaged<IntPtr, byte*, uint, uint*, int>)&Write;
-                vtblRaw[5] = (IntPtr)(delegate* unmanaged<IntPtr, long, SeekOrigin, ulong*, int>)&Seek;
-                vtblRaw[6] = (IntPtr)(delegate* unmanaged<IntPtr, ulong, int>)&SetSize;
-                vtblRaw[7] = (IntPtr)(delegate* unmanaged<IntPtr, IntPtr, ulong, ulong*, ulong*, int>)&CopyTo;
-                vtblRaw[8] = (IntPtr)(delegate* unmanaged<IntPtr, uint, int>)&Commit;
-                vtblRaw[9] = (IntPtr)(delegate* unmanaged<IntPtr, int>)&Revert;
-                vtblRaw[10] = (IntPtr)(delegate* unmanaged<IntPtr, ulong, ulong, uint, int>)&LockRegion;
-                vtblRaw[11] = (IntPtr)(delegate* unmanaged<IntPtr, ulong, ulong, uint, int>)&UnlockRegion;
-                vtblRaw[12] = (IntPtr)(delegate* unmanaged<IntPtr, Interop.Ole32.STATSTG*, Interop.Ole32.STATFLAG, int>)&Stat;
-                vtblRaw[13] = (IntPtr)(delegate* unmanaged<IntPtr, IntPtr*, int>)&Clone;
-
-                return (IntPtr)vtblRaw;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Read(IntPtr thisPtr, byte* pv, uint cb, uint* pcbRead)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.Read(pv, cb, pcbRead);
-                }
-                catch (Exception ex)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Write(IntPtr thisPtr, byte* pv, uint cb, uint* pcbWritten)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.Write(pv, cb, pcbWritten);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Seek(IntPtr thisPtr, long dlibMove, SeekOrigin dwOrigin, ulong* plibNewPosition)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.Seek(dlibMove, dwOrigin, plibNewPosition);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int SetSize(IntPtr thisPtr, ulong libNewSize)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.SetSize(libNewSize);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int CopyTo(IntPtr thisPtr, IntPtr pstm, ulong cb, ulong* pcbRead, ulong* pcbWritten)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    Interop.Ole32.IStream pstmStream = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)pstm);
-
-                    inst.CopyTo(pstmStream, cb, pcbRead, pcbWritten);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Commit(IntPtr thisPtr, uint grfCommitFlags)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.Commit((Interop.Ole32.STGC)grfCommitFlags);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Revert(IntPtr thisPtr)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.Revert();
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int LockRegion(IntPtr thisPtr, ulong libOffset, ulong cb, uint dwLockType)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    return (int)inst.LockRegion(libOffset, cb, dwLockType);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-            }
-
-            [UnmanagedCallersOnly]
-            private static int UnlockRegion(IntPtr thisPtr, ulong libOffset, ulong cb, uint dwLockType)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    return (int)inst.UnlockRegion(libOffset, cb, dwLockType);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Stat(IntPtr thisPtr, Interop.Ole32.STATSTG* pstatstg, Interop.Ole32.STATFLAG grfStatFlag)
-            {
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-                    inst.Stat(out *pstatstg, grfStatFlag);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-
-            [UnmanagedCallersOnly]
-            private static int Clone(IntPtr thisPtr, IntPtr* ppstm)
-            {
-                if (ppstm == null)
-                {
-                    return (int)Interop.HRESULT.STG_E_INVALIDPOINTER;
-                }
-
-                try
-                {
-                    Interop.Ole32.IStream inst = ComInterfaceDispatch.GetInstance<Interop.Ole32.IStream>((ComInterfaceDispatch*)thisPtr);
-
-                    *ppstm = Instance.GetOrCreateComInterfaceForObject(inst.Clone(), CreateComInterfaceFlags.None);
-                }
-                catch (Exception e)
-                {
-                    return e.HResult;
-                }
-
-                return S_OK;
-            }
-        }
-
-        private class PictureWrapper : Ole32.IPicture, Ole32.IPersistStream
-        {
-            private readonly IntPtr _wrappedInstance;
-
-            public PictureWrapper(IntPtr wrappedInstance)
-            {
-                _wrappedInstance = wrappedInstance;
-            }
-
-            public void Dispose()
-            {
-                Marshal.Release(_wrappedInstance);
-            }
-
-            public int Handle
-            {
-                get
-                {
-                    int retVal;
-                    int errorCode = ((delegate* unmanaged<IntPtr, int*, int>)(*(*(void***)_wrappedInstance + 3 /* IPicture.get_Handle slot */)))
-                        (_wrappedInstance, &retVal);
-                    if (errorCode < 0)
-                    {
-                        Marshal.ThrowExceptionForHR(errorCode);
-                    }
-
-                    return retVal;
-                }
-            }
-
-            public int hPal
-            {
-                get
-                {
-                    int retVal;
-                    int errorCode = ((delegate* unmanaged<IntPtr, int*, int>)(*(*(void***)_wrappedInstance + 4 /* IPicture.get_hPal slot */)))
-                        (_wrappedInstance, &retVal);
-                    if (errorCode < 0)
-                    {
-                        Marshal.ThrowExceptionForHR(errorCode);
-                    }
-
-                    return retVal;
-                }
-            }
-
-            public short Type
-            {
-                get
-                {
-                    short retVal;
-                    int errorCode = ((delegate* unmanaged<IntPtr, short*, int>)(*(*(void***)_wrappedInstance + 5 /* IPicture.get_Type slot */)))
-                        (_wrappedInstance, &retVal);
-                    if (errorCode < 0)
-                    {
-                        Marshal.ThrowExceptionForHR(errorCode);
-                    }
-
-                    return retVal;
-                }
-            }
-
-            public int Width
-            {
-                get
-                {
-                    int retVal;
-                    int errorCode = ((delegate* unmanaged<IntPtr, int*, int>)(*(*(void***)_wrappedInstance + 6 /* IPicture.get_Width slot */)))
-                        (_wrappedInstance, &retVal);
-                    if (errorCode < 0)
-                    {
-                        Marshal.ThrowExceptionForHR(errorCode);
-                    }
-
-                    return retVal;
-                }
-            }
-
-            public int Height
-            {
-                get
-                {
-                    int retVal;
-                    int errorCode = ((delegate* unmanaged<IntPtr, int*, int>)(*(*(void***)_wrappedInstance + 7 /* IPicture.get_Height slot */)))
-                        (_wrappedInstance, &retVal);
-                    if (errorCode < 0)
-                    {
-                        Marshal.ThrowExceptionForHR(errorCode);
-                    }
-
-                    return retVal;
-                }
-            }
-
-            public HRESULT GetClassID(
-                Guid* pClassID)
-            {
-                throw new NotImplementedException();
-            }
-
-            public HRESULT IsDirty()
-            {
-                throw new NotImplementedException();
-            }
-
-            public void Load(Interop.Ole32.IStream pstm)
-            {
-                Guid streamIID = IID_IStream;
-                IntPtr streamPtr = IntPtr.Zero;
-
-                try
-                {
-                    streamPtr = WinFormsComWrappers.Instance.GetOrCreateComInterfaceForObject(pstm, CreateComInterfaceFlags.None);
-                    Marshal.ThrowExceptionForHR(((delegate* unmanaged<IntPtr, IntPtr, int>)(*(*(void***)_wrappedInstance + 5 /* IPersistStream.Load slot */)))
-                        (_wrappedInstance, streamPtr));
-                }
-                finally
-                {
-                    if (streamPtr != IntPtr.Zero)
-                    {
-                        int count = Marshal.Release(streamPtr);
-                        Debug.Assert(count == 0);
-                    }
-                }
-            }
-
-            public void Save(Interop.Ole32.IStream pstm, Interop.BOOL fClearDirty)
-            {
-                throw new NotImplementedException();
-            }
-
-            public long GetSizeMax()
-            {
-                throw new NotImplementedException();
-            }
         }
     }
 }
